LPS-180641 Change type from number condition

+++ modules/apps/site-initializer/site-initializer-extender/site-initializer-extender/src/main/java/com/liferay/site/initializer/extender/internal/BundleSiteInitializer.java
@@ -4419,7 +4419,7 @@ public class BundleSiteInitializer implements SiteInitializer {
 		else if (jsonObject.getInt("dataType") ==
 					ExpandoColumnConstants.NUMBER) {
 
-			return jsonObject.getInt("defaultValue");
+			return jsonObject.getDouble("defaultValue");
 		}
 		else if (jsonObject.getInt("dataType") ==
 					ExpandoColumnConstants.NUMBER_ARRAY) {
